files refactoring and organizing
