+++ unsocialalien-functions/functions/constants.js
@@ -1,3 +1,3 @@
 module.exports = {
-  PASSWORD_ERROR: "auth/wrong-password"
+  PASSWORD_ERROR: 'auth/wrong-password',
 };
