+++ unsocialalien-functions/functions/util/firebaseConfig.js
@@ -1,0 +1,10 @@
+module.exports = {
+  apiKey: 'AIzaSyDwJJX0qaLg9sHSzLPFF4TKsbUK5L_c1VA',
+  authDomain: 'unsocial-alien.firebaseapp.com',
+  databaseURL: 'https://unsocial-alien.firebaseio.com',
+  projectId: 'unsocial-alien',
+  storageBucket: 'unsocial-alien.appspot.com',
+  messagingSenderId: '362380376412',
+  appId: '1:362380376412:web:0a6ecea50d564d6710e999',
+  measurementId: 'G-P8DX3K4MRV',
+};
